Rename callback_interface.py to callback.py; deprecate CallbackDecorator alias
