+++ RedisBridge/interfaces/__init__.py
@@ -1,5 +1,2 @@
 from .base import RedisInterface
-from .callback_interface import CallbackInterface
-
-# Alias for backward compatibility
-CallbackDecorator = CallbackInterface
+from .callback import CallbackInterface
