+++ RedisBridge/interfaces/callback.py
